+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added an Equality comparison overload to ByteArray
-- add empty() member function to ByteArray
+- added empty() member function to ByteArray
+- added create_from_string(std::string_view) factory function
 
 ### Fixed
 - Added template restrictions for iterator constructor to avoid ambiguity between constructors